+++ src/store/taskStore.ts
@@ -1,6 +1,7 @@
 import { create } from "zustand";
 
 import taskType from "../types/taskType";
+import dummyTasks from "../utils/dummyTasks";
 
 type state = {
   tasks: taskType[];
@@ -18,7 +19,8 @@ type actions = {
 
 const useTaskStore = create<state & actions>()((set) => ({
   // Stores all task items; they are objects of type taskType
-  tasks: [],
+  // tasks: [],
+  tasks: dummyTasks,
   // Add a task
   addTask: (task: taskType) =>
     set((state) => ({