+++ src/types/taskType.ts
@@ -4,7 +4,7 @@ type taskType = {
   name: string;
   category: string;
   description: string;
-  priority: 1 | 2 | 3; // 1 is lowest priority, 3 is highest priority
+  priority: 0 | 1 | 2 | 3; // 1 is lowest priority, 3 is highest priority
   isComplete: boolean;
 };