+++ src/utils/routeSegments.ts
@@ -1,4 +1,5 @@
-// All valid individual route segments
+/*** All valid individual route segments ***/
+
 export const FRONT = "Todo-App";
 export const ADD_TASK = "add-task";
 export const EDIT_TASK = "edit-task";